# docs/VMOGD3swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "VMOGD3Project",
    "contact": {}
  },
  "host": "localhost:5000",
  "basePath": "/api/v1",
  "securityDefinitions": {
    "token": {
      "type": "apiKey",
      "description": "",
      "name": "token",
      "in": "header"
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/techs/{id}": {
      "get": {
        "summary": "Get tech by id",
        "tags": [
          "Tech"
        ],
        "operationId": "Gettechbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Update tech",
        "tags": [
          "Tech"
        ],
        "operationId": "Updatetech",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdatetechRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete tech",
        "tags": [
          "Tech"
        ],
        "operationId": "Deletetech",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/projects/{id}": {
      "get": {
        "summary": "Get project by id",
        "tags": [
          "Project"
        ],
        "operationId": "Getprojectbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "put": {
        "summary": "Update project",
        "tags": [
          "Project"
        ],
        "operationId": "Updateproject",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateprojectRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete project",
        "tags": [
          "Project"
        ],
        "operationId": "Deleteproject",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "UpdatetechRequest": {
      "title": "UpdatetechRequest",
      "example": {
        "name": "NodeJS"
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      },
      "required": [
        "name"
      ]
    },
...
    "UpdateprojectRequest": {
      "title": "UpdateprojectRequest",
      "example": {
        "name": "Big project",
        "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Orci a scelerisque purus semper.",
        "type": "dee61708-4182-408e-bf0e-4b3bdc527eba",
        "status": "Maintain",
        "customer": "5b119ad0-4ae2-4c7f-b815-cb8395838939",
        "startDate": "05/21/2021",
        "teches": [
          "9bbe5a2e-e8f8-4593-8ead-2b12b1ff72e1",
          "215a177f-4307-4cf3-a036-d0d0a0a41eff",
          "44685a01-1549-4a59-832d-7964ec86f648",
          "3ed0ab8d-204d-4803-8586-d521822d9268"
        ],
        "employees": [
          "b5a0e04d-6bcc-4100-b3cc-17a5811882ef",
          "4841b4dc-9825-4a34-8e62-b40159beef20",
          "559a479d-99c2-4441-95d4-6ab00b44c7a4",
          "8efb0d42-5d65-43ae-83cc-89d6e3bc6d31"
        ]
      },
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "customer": {
          "type": "string"
        },
        "startDate": {
          "type": "string"
        },
        "teches": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "employees": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "name",
        "description",
        "type",
        "status",
        "customer",
        "startDate",
        "teches",
        "employees"
      ]
    }
  },
  "security": [
    {
      "token": []
    }
  ],
...
}